Updating Superset image and documentation
Updating the Superset image in the docker-compose.yml to use the
amancevice/superset image which is the one recommended in the official
Superset documentation.

This image is much more up to date and doesn't suffer some of the bugs
the previous one had.

Now an extra step is required to initialize the superset database and
create the admin user.